requirements.txt
altgraph==0.17.3
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
branca==0.6.0
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==3.0.1
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
docutils==0.19
executing==1.2.0
Fiona==1.9.1
folium==0.14.0
fonttools==4.38.0
geographiclib==2.0
geopandas==0.12.2
geopy==2.3.0
idna==3.4
ipython==8.11.0
jedi==0.18.2
Jinja2==3.1.2
Kivy==2.1.0
Kivy-Garden==0.1.5
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mplleaflet==0.0.5
munch==2.5.0
numpy==1.24.2
packaging==23.0
pandas==1.5.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.14.0
pyinstaller==5.8.0
pyinstaller-hooks-contrib==2023.0
pyparsing==3.0.9
pyproj==3.4.1
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
shapely==2.0.1
six==1.16.0
stack-data==0.6.2
tk==0.1.0
traitlets==5.9.0
urllib3==1.26.14
wcwidth==0.2.6
